requirements.txt
streamlit==1.50.0
pandas==2.3.1
google-generativeai
python-dotenv
matplotlib